+++ main/src/main/java/com/github/topisenpai/lavasrc/mirror/MirroringAudioTrack.java
@@ -1,24 +1,15 @@
 package com.github.topisenpai.lavasrc.mirror;
 
-import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler;
 import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
 import com.sedmelluq.discord.lavaplayer.track.AudioItem;
 import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist;
-import com.sedmelluq.discord.lavaplayer.track.AudioReference;
-import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
 import com.sedmelluq.discord.lavaplayer.track.DelegatedAudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.InternalAudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.playback.LocalAudioTrackExecutor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.CompletableFuture;
 
 public abstract class MirroringAudioTrack extends DelegatedAudioTrack {
-
-    private static final Logger log = LoggerFactory.getLogger(MirroringAudioTrack.class);
 
     protected final String isrc;
     protected final String artworkURL;
@@ -37,14 +28,6 @@ public abstract class MirroringAudioTrack extends DelegatedAudioTrack {
 
     public String getArtworkURL() {
         return this.artworkURL;
-    }
-
-    public String getTrackTitle() {
-        var query = this.trackInfo.title;
-        if (!this.trackInfo.author.equals("unknown")) {
-            query += " " + this.trackInfo.author;
-        }
-        return query;
     }
 
     @Override
@@ -68,37 +51,6 @@ public abstract class MirroringAudioTrack extends DelegatedAudioTrack {
 
     public MirroringAudioSourceManager getMirroringSourceManager() {
         return this.sourceManager;
-    }
-
-    public static AudioItem loadItem(String query, MirroringAudioSourceManager sourceManager) {
-        var cf = new CompletableFuture<AudioItem>();
-        sourceManager.getAudioPlayerManager().loadItem(query, new AudioLoadResultHandler() {
-
-            @Override
-            public void trackLoaded(AudioTrack track) {
-                log.debug("Track loaded: " + track.getIdentifier());
-                cf.complete(track);
-            }
-
-            @Override
-            public void playlistLoaded(AudioPlaylist playlist) {
-                log.debug("Playlist loaded: " + playlist.getName());
-                cf.complete(playlist);
-            }
-
-            @Override
-            public void noMatches() {
-                log.debug("No matches found for: " + query);
-                cf.complete(AudioReference.NO_TRACK);
-            }
-
-            @Override
-            public void loadFailed(FriendlyException exception) {
-                log.debug("Failed to load: " + query);
-                cf.completeExceptionally(exception);
-            }
-        });
-        return cf.join();
     }
 
 }
